Fix gathering horn failing in multiplayer

+++ Items/BuddyHorn.cs
@@ -43,37 +43,40 @@ namespace WeaponOut.Items
 
         public override bool UseItem(Player player)
         {
-            int offsetY = (int)(-11 * player.gravDir);
-            
-            int spawnX = Main.spawnTileX;
-            int spawnY = Main.spawnTileY;
-            player.FindSpawn();
-            if (player.SpawnX > 0) spawnX = player.SpawnX;
-            if (player.SpawnY > 0) spawnY = player.SpawnY;
-            
-            int entrance = mod.ProjectileType<Projectiles.BuddyPortalEntrance>();
-            int exit = mod.ProjectileType<Projectiles.BuddyPortalExit>();
-
-            foreach (Projectile projectile in Main.projectile)
+            if (player.whoAmI == Main.myPlayer)
             {
-                if (projectile.active && projectile.owner == player.whoAmI)
-                {
-                    if (projectile.type == entrance) projectile.Kill();
-                    if (projectile.type == exit) projectile.Kill();
-                }
-            }
+                int offsetY = (int)(-11 * player.gravDir);
 
-            // Entrance
-            Projectile.NewProjectile(new Vector2(
-                spawnX, spawnY).ToWorldCoordinates() - new Vector2(0, 40),
-                default(Vector2),
-                entrance, 0, 0f, player.whoAmI);
-            // Exit
-            Projectile.NewProjectile(new Vector2(
-                (int)player.Center.X,
-                (int)player.Center.Y + offsetY
-                ), default(Vector2),
-                exit, 0, 0f, player.whoAmI);
+                int spawnX = Main.spawnTileX;
+                int spawnY = Main.spawnTileY;
+                player.FindSpawn();
+                if (player.SpawnX > 0) spawnX = player.SpawnX;
+                if (player.SpawnY > 0) spawnY = player.SpawnY;
+
+                int entrance = mod.ProjectileType<Projectiles.BuddyPortalEntrance>();
+                int exit = mod.ProjectileType<Projectiles.BuddyPortalExit>();
+
+                foreach (Projectile projectile in Main.projectile)
+                {
+                    if (projectile.active && projectile.owner == player.whoAmI)
+                    {
+                        if (projectile.type == entrance) projectile.Kill();
+                        if (projectile.type == exit) projectile.Kill();
+                    }
+                }
+
+                // Entrance
+                Projectile.NewProjectile(new Vector2(
+                    spawnX, spawnY).ToWorldCoordinates() - new Vector2(0, 40),
+                    default(Vector2),
+                    entrance, 0, 0f, player.whoAmI);
+                // Exit
+                Projectile.NewProjectile(new Vector2(
+                    (int)player.Center.X,
+                    (int)player.Center.Y + offsetY
+                    ), default(Vector2),
+                    exit, 0, 0f, player.whoAmI);
+            }
             return true;
         }
     }
